step 3 - is this how it is supposed to work?
